semana-08/rinha-backend/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg 
pydantic
python-dotenv  
ormar
databases[postgresql]
gunicorn
psycopg2-binary
